+++ src/main/java/mg/tommy/springboot/springbootwebapp/service/library/CustomerServiceImpl.java
@@ -11,6 +11,7 @@ import org.springframework.util.StringUtils;
 import java.time.LocalDateTime;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.concurrent.atomic.AtomicReference;
 
 import static java.util.stream.Collectors.toList;
 
@@ -44,11 +45,10 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
-    public CustomerDto overwriteById(UUID uuid, CustomerDto customerDto) {
-        Customer customer = customerMapper.toCustomer(customerDto);
+    public Optional<CustomerDto> overwriteById(UUID uuid, CustomerDto customerDto) {
         Optional<Customer> customerToUpdate = customerRepository.findById(uuid);
         if (customerToUpdate.isEmpty()) {
-            return customerMapper.toCustomerDto(customerRepository.save(customer));
+            return Optional.empty();
         }
         Customer.CustomerBuilder builder = customerToUpdate.get().toBuilder()
                 .firstName(customerDto.getFirstName())
@@ -56,34 +56,41 @@ public class CustomerServiceImpl implements CustomerService {
                 .email(customerDto.getEmail())
                 .birthdate(customerDto.getBirthdate())
                 .updateDate(LocalDateTime.now());
-        return customerMapper.toCustomerDto(customerRepository.save(builder.build()));
+        return Optional.of(customerMapper.toCustomerDto(customerRepository.save(builder.build())));
     }
 
     @Override
-    public CustomerDto updateById(UUID uuid, CustomerDto customerDto) {
-        Optional<Customer> customerToUpdate = customerRepository.findById(uuid);
-        if (customerToUpdate.isEmpty()) {
-            return null;
-        }
-        Customer.CustomerBuilder builder = customerToUpdate.get().toBuilder();
-        if (StringUtils.hasText(customerDto.getFirstName())) {
-            builder.firstName(customerDto.getFirstName());
-        }
-        if (StringUtils.hasText(customerDto.getLastName())) {
-            builder.lastName(customerDto.getLastName());
-        }
-        if (StringUtils.hasText(customerDto.getEmail())) {
-            builder.email(customerDto.getEmail());
-        }
-        if (customerDto.getBirthdate() != null) {
-            builder.birthdate(customerDto.getBirthdate());
-        }
-        builder.updateDate(LocalDateTime.now());
-        return customerMapper.toCustomerDto(customerRepository.save(builder.build()));
+    public Optional<CustomerDto> updateById(UUID uuid, CustomerDto customerDto) {
+        AtomicReference<Optional<CustomerDto>> customerDtoReference = new AtomicReference<>();
+
+        customerRepository.findById(uuid)
+                .ifPresentOrElse(customerToUpdate -> {
+                    Customer.CustomerBuilder builder = customerToUpdate.toBuilder().updateDate(LocalDateTime.now());
+                    if (StringUtils.hasText(customerDto.getFirstName())) {
+                        builder.firstName(customerDto.getFirstName());
+                    }
+                    if (StringUtils.hasText(customerDto.getLastName())) {
+                        builder.lastName(customerDto.getLastName());
+                    }
+                    if (StringUtils.hasText(customerDto.getEmail())) {
+                        builder.email(customerDto.getEmail());
+                    }
+                    if (customerDto.getBirthdate() != null) {
+                        builder.birthdate(customerDto.getBirthdate());
+                    }
+                    Customer updatedCustomer = customerRepository.save(builder.build());
+                    customerDtoReference.set(Optional.of(customerMapper.toCustomerDto(updatedCustomer)));
+                }, () -> customerDtoReference.set(Optional.empty()));
+
+        return customerDtoReference.get();
+
     }
 
     @Override
-    public void deleteById(UUID uuid) {
+    public boolean deleteById(UUID uuid) {
+        if (!customerRepository.existsById(uuid))
+            return false;
         customerRepository.deleteById(uuid);
+        return true;
     }
 }
